+++ Documentation/WhatsNew.md
@@ -11,6 +11,7 @@ Milestone 4
   * Added ezDynamicOctree and ezDynamicQuadtree to the CoreUtils library, which are data structures to optimize range queries.
   * Added Lua to the ThirdParty library.
   * Added ezHashedString and ezTempHashedString for storing strings that rarely change but where comparisons should be really fast.
+  * Added ezFrustum for object culling.
 
 Milestone 3
 -----------